requirements.txt
absl-py==1.3.0
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
cloudpickle==2.2.0
Cython==0.29.32
decorator==4.4.2
dm-control==1.0.9
dm-env==1.5
dm-haiku==0.0.9
dm-tree==0.1.8
exceptiongroup==1.0.4
fasteners==0.18
flatbuffers==22.12.6
gast==0.4.0
glfw==2.5.5
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
importlib-metadata==5.2.0
iniconfig==1.1.1
jax==0.4.1
-e git+ssh://git@github.com/lasgroup/jax-cpo.git@100d365b9cd42f064d460680d03423b29fc38f09#egg=jax_cpo
jaxlib==0.4.1
jmp==0.0.2
keras==2.11.0
labmaze==1.0.6
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
moviepy==1.0.3
mujoco==2.3.1.post1
mujoco-py==2.1.2.14
numpy==1.24.0
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.4
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safe-adaptation-gym @ https://github.com/lasgroup/safe-adaptation-gym/tarball/main
scipy==1.9.3
six==1.16.0
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
tensorflow-probability==0.19.0
termcolor==2.1.1
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.11.0
